[go-gen] Refactor DAO generation into smaller files
