+++ src/main/scala/temple/generate/server/go/service/GoServiceGenerator.scala
@@ -2,6 +2,7 @@ package temple.generate.server.go.service
 
 import temple.generate.FileSystem._
 import temple.generate.server.go.common._
+import temple.generate.server.go.service.dao.GoServiceDAOGenerator
 import temple.generate.server.{ServiceGenerator, ServiceRoot}
 import temple.generate.utils.CodeTerm.mkCode
